requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.26.37
botocore==1.29.37
Django==4.1.4
django-appconf==1.0.5
django-imagekit==4.1.0
django-storages==1.13.2
jmespath==1.0.1
pilkit==2.0
Pillow==9.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
s3transfer==0.6.0
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
urllib3==1.26.13
